Refactor v1 operator dispatcher
