+++ crates/rulemorph/src/transform/operators.rs
@@ -4,6 +4,7 @@ mod args;
 mod array;
 mod boolean;
 mod date;
+mod dispatch;
 mod json;
 mod lookup;
 mod number;
@@ -16,589 +17,13 @@ pub(super) use self::args::{arg_expr_at, args_len};
 use self::args::{
     eval_arg_string_at, eval_arg_value_at, eval_expr_at_index, eval_expr_value_or_null_at,
 };
-use self::array::{
-    eval_array_avg, eval_array_chunk, eval_array_contains, eval_array_distinct_by, eval_array_drop,
-    eval_array_filter, eval_array_find, eval_array_find_index, eval_array_flat_map,
-    eval_array_flatten, eval_array_fold, eval_array_group_by, eval_array_index_of,
-    eval_array_key_by, eval_array_map, eval_array_max, eval_array_min, eval_array_partition,
-    eval_array_reduce, eval_array_slice, eval_array_sort_by, eval_array_sum, eval_array_take,
-    eval_array_unique, eval_array_unzip, eval_array_zip, eval_array_zip_with,
-};
-use self::boolean::{compare_eq, eval_bool_and_or, eval_bool_not, eval_compare};
-use self::date::{eval_date_format, eval_to_unixtime};
-use self::json::{
-    eval_json_entries, eval_json_from_entries, eval_json_get, eval_json_keys, eval_json_merge,
-    eval_json_object_flatten, eval_json_object_unflatten, eval_json_omit, eval_json_pick,
-    eval_json_values, eval_len,
-};
-use self::lookup::eval_lookup;
-use self::number::{eval_numeric_op, eval_round, eval_to_base};
+use self::boolean::compare_eq;
+pub(crate) use self::dispatch::eval_op;
 pub(super) use self::shared::{
     SortKey, SortKeyKind, compare_sort_keys, locals_with_item, locals_with_precomputed_args,
 };
-use self::simple::{eval_coalesce, eval_concat};
-use self::string::{eval_pad, eval_replace, eval_split, eval_unary_string_op};
 pub(super) use self::value::{cast_value, value_as_bool, value_to_string};
 use self::value::{
     expr_type_error, json_number_from_f64, to_radix_string, value_as_string, value_to_i64,
     value_to_number, value_to_string_optional,
 };
-
-pub(crate) fn eval_op(
-    expr_op: &ExprOp,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    injected: Option<&EvalValue>,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(&expr_op.args, injected);
-    if total_len == 0 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must be a non-empty array",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    match expr_op.op.as_str() {
-        "concat" => eval_concat(expr_op, injected, record, context, out, base_path, locals),
-        "coalesce" => eval_coalesce(expr_op, injected, record, context, out, base_path, locals),
-        "to_string" => eval_unary_string_op(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-            |value, path| value_to_string(value, path).map(JsonValue::String),
-        ),
-        "trim" => eval_unary_string_op(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-            |value, path| {
-                let s = value_as_string(value, path)?;
-                Ok(JsonValue::String(s.trim().to_string()))
-            },
-        ),
-        "lowercase" => eval_unary_string_op(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-            |value, path| {
-                let s = value_as_string(value, path)?;
-                Ok(JsonValue::String(s.to_lowercase()))
-            },
-        ),
-        "uppercase" => eval_unary_string_op(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-            |value, path| {
-                let s = value_as_string(value, path)?;
-                Ok(JsonValue::String(s.to_uppercase()))
-            },
-        ),
-        "replace" => eval_replace(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "split" => eval_split(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "pad_start" => eval_pad(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            true,
-            locals,
-        ),
-        "pad_end" => eval_pad(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            false,
-            locals,
-        ),
-        "lookup" => eval_lookup(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            false,
-            locals,
-        ),
-        "lookup_first" => eval_lookup(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            true,
-            locals,
-        ),
-        "merge" => eval_json_merge(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            false,
-            locals,
-        ),
-        "deep_merge" => eval_json_merge(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            true,
-            locals,
-        ),
-        "get" => eval_json_get(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "pick" => eval_json_pick(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "omit" => eval_json_omit(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "keys" => eval_json_keys(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "values" => eval_json_values(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "entries" => eval_json_entries(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "len" => eval_len(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "from_entries" => eval_json_from_entries(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "object_flatten" => eval_json_object_flatten(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "object_unflatten" => eval_json_object_unflatten(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "map" => eval_array_map(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "filter" => eval_array_filter(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "flat_map" => eval_array_flat_map(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "flatten" => eval_array_flatten(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "take" => eval_array_take(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "drop" => eval_array_drop(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "slice" => eval_array_slice(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "chunk" => eval_array_chunk(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "zip" => eval_array_zip(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "zip_with" => eval_array_zip_with(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "unzip" => eval_array_unzip(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "group_by" => eval_array_group_by(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "key_by" => eval_array_key_by(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "partition" => eval_array_partition(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "unique" => eval_array_unique(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "distinct_by" => eval_array_distinct_by(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "sort_by" => eval_array_sort_by(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "find" => eval_array_find(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "find_index" => eval_array_find_index(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "index_of" => eval_array_index_of(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "contains" => eval_array_contains(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "sum" => eval_array_sum(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "avg" => eval_array_avg(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "min" => eval_array_min(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "max" => eval_array_max(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "reduce" => eval_array_reduce(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "fold" => eval_array_fold(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "+" | "-" | "*" | "/" => {
-            eval_numeric_op(expr_op, injected, record, context, out, base_path, locals)
-        }
-        "round" => eval_round(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "to_base" => eval_to_base(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "date_format" => eval_date_format(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "to_unixtime" => eval_to_unixtime(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "and" => eval_bool_and_or(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            true,
-            locals,
-        ),
-        "or" => eval_bool_and_or(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            false,
-            locals,
-        ),
-        "not" => eval_bool_not(
-            &expr_op.args,
-            injected,
-            record,
-            context,
-            out,
-            base_path,
-            locals,
-        ),
-        "==" | "!=" | "<" | "<=" | ">" | ">=" | "~=" => {
-            eval_compare(expr_op, injected, record, context, out, base_path, locals)
-        }
-        _ => Err(
-            TransformError::new(TransformErrorKind::ExprError, "expr.op is not supported")
-                .with_path(format!("{}.op", base_path)),
-        ),
-    }
-}
